Add code to check feed and check which episodes are missing
The RSS feed does not seem to support If-Modified-Since but include it
anyway just in case it works in the future or a bug is found in the
code.

Read the feed, determine which episodes are present and have a full
description, then check the wiki to determine which episodes do not yet
have a wiki page. Every episode has a redirect, episode number redirects
to the full title, so check the redirect to determine if the wiki page
exists. For now, print out the wiki text to check this works in a cron.