Merge using local feed and checking into single flow
Make --test be a boolean option to enable test mode. Add the --rss
option to take the RSS XML file as input. This allows using a local file
but also make changes, such as to backfill episodes with missing wiki
pages. To aid backfilling, add the --all option to check every single
episode with a description in the XML file.